feat: RA-78 Refactor PostgresLedgerDataProviderService and extract block* methods to LedgerBlockService
