+++ api/src/main/java/org/cardanofoundation/rosetta/api/block/service/LedgerBlockService.java
@@ -1,0 +1,33 @@
+package org.cardanofoundation.rosetta.api.block.service;
+
+import java.util.List;
+
+import org.cardanofoundation.rosetta.api.block.model.domain.Block;
+import org.cardanofoundation.rosetta.api.block.model.domain.BlockTx;
+import org.cardanofoundation.rosetta.api.block.model.domain.GenesisBlock;
+
+public interface LedgerBlockService {
+
+  /**
+   * Returns a block by its number and hash. Including all populated Transactions.
+   *
+   * @param number block number
+   * @param hash   block hash
+   * @return the block
+   */
+  Block findBlock(Long number, String hash);
+
+
+  /**
+   * Returns a list of all transactions within a block. The UTXO aren't populated yet. They contain only the hash and the index.
+   * @param number block number
+   * @param hash block hash
+   * @return the list of transactions
+   */
+  List<BlockTx> findTransactionsByBlock(Long number, String hash);
+
+
+  Block findLatestBlock();
+
+  GenesisBlock findGenesisBlock();
+}
