+++ api/src/main/java/org/cardanofoundation/rosetta/common/services/LedgerDataProviderService.java
@@ -18,35 +18,10 @@ import org.cardanofoundation.rosetta.api.block.model.domain.StakeAddressBalance;
  */
 public interface LedgerDataProviderService {
 
-  GenesisBlock findGenesisBlock();
-
-  /**
-   * Returns a block by its number and hash. Including all populated Transactions.
-   *
-   * @param number block number
-   * @param hash   block hash
-   * @return the block
-   */
-  Block findBlock(Long number, String hash);
-
   List<AddressBalance> findBalanceByAddressAndBlock(String address, Long number);
 
   List<Utxo> findUtxoByAddressAndCurrency(String address, List<Currency> currencies);
 
   List<StakeAddressBalance> findStakeAddressBalanceByAddressAndBlock(String address, Long number);
-
-  Long findLatestBlockNumber();
-
-  Block findLatestBlock();
-
-  /**
-   * Returns a list of all transactions within a block. The UTXO aren't populated yet. They contain only the hash and the index.
-   * @param number block number
-   * @param hash block hash
-   * @return the list of transactions
-   */
-  List<BlockTx> findTransactionsByBlock(Long number, String hash);
-
-  ProtocolParams findProtocolParametersFromIndexerAndConfig();
 
 }
