+++ api/src/main/java/org/cardanofoundation/rosetta/common/services/ProtocolParamService.java
@@ -5,4 +5,6 @@ import org.cardanofoundation.rosetta.api.block.model.domain.ProtocolParams;
 
 public interface ProtocolParamService {
     ProtocolParams getProtocolParameters();
+
+    ProtocolParams findProtocolParametersFromIndexerAndConfig();
 }
