+++ api/src/test/java/org/cardanofoundation/rosetta/common/services/impl/ProtocolParamServiceImplTest.java
@@ -3,15 +3,29 @@ package org.cardanofoundation.rosetta.common.services.impl;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+import org.cardanofoundation.rosetta.api.block.model.repository.EpochParamRepository;
+import org.cardanofoundation.rosetta.common.mapper.ProtocolParamsToEntity;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+@ExtendWith(MockitoExtension.class)
 class ProtocolParamServiceImplTest {
 
   ObjectMapper objectMapper = new ObjectMapper();
-  ProtocolParamServiceImpl genesisService = new ProtocolParamServiceImpl(objectMapper);
+
+  @Mock
+  EpochParamRepository epochParamRepository;
+  @Mock
+  ProtocolParamsToEntity protocolParamsToEntity;
+
+  ProtocolParamServiceImpl genesisService =
+      new ProtocolParamServiceImpl(objectMapper,epochParamRepository,protocolParamsToEntity);
 
   @Test
   void getProtocolParameters() {
